% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{addItemToObject}
\alias{addItemToObject}
\title{Adds new item to a repseq object}
\usage{
addItemToObject(object, item, itemName)
}
\arguments{
\item{object}{is a repseq object containing all repertoire sample data}

\item{item}{is the item to be added to the repseq object}

\item{itemName}{the name of the new item to be added to the repseq object}
}
\value{
a repSeq object with the new item added to its list of members.
}
\description{
Adds new item to a repseq object
}
