% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CDRepseqObj}
\alias{CDRepseqObj}
\title{Celiac disease (CD) PBMC Repseq data object}
\format{A list with 3 elements: \code{sampleData}, \code{samNames} and
  \code{group}}
\usage{
CDRepseqObj
}
\description{
The CD PBMC Repseq data object is a list with three elements. First element contains the 
list of 8 immune repertoire samples, second element contains a vector of the names of the 
samples, third element contains a vector of their condition status in zeros and ones.
The object is prepared by the setUp function as described in the help of setUp or on RepAn github ReadMe.
}
\keyword{datasets}
