% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{compareAbundanceInPairedSamplesForRanking}
\alias{compareAbundanceInPairedSamplesForRanking}
\title{perform fisher's exact test based ranking of candidate DA CDR3s}
\usage{
compareAbundanceInPairedSamplesForRanking(samObj, freqTable,
  pairs = NULL, paired = T)
}
\description{
perform fisher's exact test based ranking of candidate DA CDR3s
}
\keyword{internal}
