% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{compareClusterAbundancesPaired}
\alias{compareClusterAbundancesPaired}
\title{compare cluster abundances for paired cases}
\usage{
compareClusterAbundancesPaired(sam1, sam2, clusMatchTable, s1cls, s2cls)
}
\description{
compare cluster abundances for paired cases
}
