% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{compareClusterAbundancesUnPaired}
\alias{compareClusterAbundancesUnPaired}
\title{compare cluster abundances for unpaired cases}
\usage{
compareClusterAbundancesUnPaired(repSeqObj)
}
\description{
compare cluster abundances for unpaired cases
}
\keyword{internal}
