% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{computeVgeneEnrichement}
\alias{computeVgeneEnrichement}
\title{V-gene usage analysis in differentially abundant CDR3s}
\usage{
computeVgeneEnrichement(repSamObj, DaClonotypes, n = 20)
}
\arguments{
\item{repSamObj}{is a repseq data object containing all repertoire sample data}

\item{DaClonotypes}{a vector of CDR3 amino acid sequences that are considered to be differentially abundant}

\item{n}{the number of repeat resamples to perform; each random resample samples CDR3s the same size as the DaClonotypes. Default is 20}
}
\value{
function returns a data frame with the observed frequency, frequencies in random resamples, mean of the frequencies in the random resamples,
the p-value estimated from the null distribution of resampled frequencies, and percent increase in v-gene usage for every V-genes used in the DA DR3s
}
\description{
This function compares observed V-gene frequencies to frequencies obtained from randomly sampled sets of CDR3s from pooled repertoires of all samples to evaluate bias in V-gene usage
}
