% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{CountkmerFrequency}
\alias{CountkmerFrequency}
\title{Count kmer frequencies in a CDR3}
\usage{
CountkmerFrequency(seqs, type = "NT", k = 4)
}
\arguments{
\item{seqs}{a vector of all CDR3 sequences}

\item{type}{the type of kmers, NT or AA}

\item{k}{the size of k, default is 4}
}
\value{
data frame of sequences versus kmer counts
}
\description{
Count kmer frequencies in a CDR3
}
