% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{createDecoyCDR3s_V2}
\alias{createDecoyCDR3s_V2}
\title{Create decoy CDR3 sequences and add to each repertoire sample}
\usage{
createDecoyCDR3s_V2(repSeqObj, seqType = c("NT", "AA"),
  decoyProp = 0.5, randromSelect = T, paired = T,
  referenceSetFileInRDS = NULL)
}
\arguments{
\item{repSeqObj}{is an object containing all repertoire sample data}

\item{seqType}{subsequence type to consider, either NT (nucleotide) or AA (amino acid).}

\item{decoyProp}{the proportion of decoy CDR3s to include in each sample (this proportion is in comparison to the resampled repertoire size for each sample). Default is 0.5.}

\item{randromSelect}{Boolean; should the decoy CDR3s be randomly sampled from reference or in-silico generated. Default is random sampling from the reference.}

\item{paired}{Boolean; whether samples are matched/paired.If true, and randromSelect is true, random selection of decoys for the matched samples is done from a random subset of CDR3s from the total reference.}

\item{referenceSetFileInRDS}{Reference TCR CDR3 repertoire (in immunoseq format and in RDS file type) to use as a reference set. If not provided, the reference dataset would be used that comes with the package would be used.}
}
\value{
repSeq data Object containing repertoire samples that contain decoy CDR3 sequences
}
\description{
Function creates decoy CDR3 sequences either by random sampling from a reference repertoire or by in-silico generation per each sample
}
