% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{determineWeight}
\alias{determineWeight}
\title{Find starting position of kmer cdr3 sequence
Determine weight of kmer depending on position}
\usage{
determineWeight(kmer, seq)
}
\description{
Find starting position of kmer cdr3 sequence
Determine weight of kmer depending on position
}
