% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{extractDASubRepertoire}
\alias{extractDASubRepertoire}
\title{Extract DA subrepertoire and CDR3s contained in them}
\usage{
extractDASubRepertoire(repSeqObj, cutoff = 0.1, method = "pvalue")
}
\description{
Extract DA subrepertoire and CDR3s contained in them
}
