% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{extractSubRepertoire}
\alias{extractSubRepertoire}
\title{Extract any subrepertoire and CDR3s contained in them}
\usage{
extractSubRepertoire(repSeqObj, subReps = NULL)
}
\description{
Extract any subrepertoire and CDR3s contained in them
}
