% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{findDAClonotypesInDACluster}
\alias{findDAClonotypesInDACluster}
\title{Find out previously known Celiac disease CDR3s among DA clonotypes detected}
\usage{
findDAClonotypesInDACluster(clonotypes)
}
\description{
Find out previously known Celiac disease CDR3s among DA clonotypes detected
}
