% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{findDAClusters}
\alias{findDAClusters}
\title{Find differentially abundant subrepertoires}
\usage{
findDAClusters(repSeqObj, abundanceType = c("cAbundance",
  "cRelAbundance", "cRelCloneSize"), testType = c("t.test",
  "wilcox.test", "RankProd"), paired = F, ...)
}
\description{
Find differentially abundant subrepertoires
}
\keyword{internal}
