% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{findOptimalClusters}
\alias{findOptimalClusters}
\title{finds optimal clusters within samples and matches them across samples}
\usage{
findOptimalClusters(repSeqObj, k, clusterby = "NT", kmerWidth = 4,
  posWt = F, distMethod = "euclidean", useDynamicTreeCut = T,
  matchingMethod = c("hc", "km", "og"))
}
\description{
finds optimal clusters within samples and matches them across samples
}
