% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{findOptimalK}
\alias{findOptimalK}
\title{Finding optimal number of clusters}
\usage{
findOptimalK(repSeqObj, nSamEval = 2, clusterby, minCSizePerc = 0.1,
  minNClonesPerCluster = 20, kmerWidth = 4, posWt = T,
  distMethod = "euclidean")
}
\arguments{
\item{repSeqObj}{is an object containing all repertoire sample data}

\item{posWt}{boolean to give weights to kmer frequencies depending on their position in the CDR3}

\item{distMethod}{the distance method used for determining distance between CDR3 feature vectors, default "euclidean"}
}
\value{
returns an optimal k for dividing unsupervised clustering results into k compact clusters.
}
\description{
Finds optimal k as the average optimal k detected from the within sample clustering of selected repertoire samples.
}
