% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{getClusterAbundancesTable}
\alias{getClusterAbundancesTable}
\title{estimate the abundance of cluster of CDR3s}
\usage{
getClusterAbundancesTable(repSeqObj)
}
\description{
estimate the abundance of cluster of CDR3s
}
\keyword{internal}
