% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{getClusterFoldChanges}
\alias{getClusterFoldChanges}
\title{subrepertoire fold change}
\usage{
getClusterFoldChanges(sam1, sam2, consensusT, s1cls, s2cls)
}
\description{
subrepertoire fold change
}
\keyword{internal}
