% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{getClusterLables}
\alias{getClusterLables}
\title{This function performs pairwise 4-mer nt usage clustering of clonotypes in repSeq data}
\usage{
getClusterLables(sam, k = 10, clusterby = "NT", kmerWidth = 4,
  posWt = F, distMethod = "euclidean", useDynamicTreeCut = T)
}
\arguments{
\item{sam}{a repertoire data (of only productive clonotypes,immunoseq format) for whose clonotypes are going to be clustered based on 4-mer usage frequencies.}
}
\value{
a list containing the clonotypes,clonotypes with 4-mer usage frequencies,distance matrix between clonotypes,cluster labels for each clonotype,and cluster centroids.
}
\description{
This function performs pairwise 4-mer nt usage clustering of clonotypes in repSeq data
}
