% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{getClusterMatches}
\alias{getClusterMatches}
\title{find matching clusters of CDR3s}
\usage{
getClusterMatches(repSeqObj, matchingMethod = c("hc", "km", "og"),
  distMethod = "euclidean")
}
\description{
find matching clusters of CDR3s
}
