% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepAnFns.R
\name{getDefined}
\alias{getDefined}
\title{Get CDR3s that have well defined V gene name from a repertoire data (Repseq sample) in immunoseq format}
\usage{
getDefined(rep)
}
\arguments{
\item{rep}{name of the repertoire sample (which has already been read in with for example with read.table)}
}
\value{
a subset of the repertoire (rep) containing CDR3s with well defined v gene name is returned
}
\description{
Get CDR3s that have well defined V gene name from a repertoire data (Repseq sample) in immunoseq format
}
