% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepAnFns.R
\name{getHasStop}
\alias{getHasStop}
\title{Get stop codon containing CDR3s from a repertoire data (Repseq sample) in immunoseq format}
\usage{
getHasStop(rep)
}
\arguments{
\item{rep}{name of the repertoire sample (which has already been read in with for example with read.table)}
}
\value{
a subset of the repertoire (rep) containing CDR3s with stop codon is returned
}
\description{
Get stop codon containing CDR3s from a repertoire data (Repseq sample) in immunoseq format
}
