% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{getKmerFrequency}
\alias{getKmerFrequency}
\title{Count kmer frequencies in a CDR3}
\usage{
getKmerFrequency(seqs, type = "NT", k = 4, normForLength = F)
}
\arguments{
\item{seqs}{a vector of all CDR3 sequences}

\item{type}{the type of kmers, NT or AA}

\item{k}{the size of k, default is 4}

\item{normForLength}{normalize the kmer frequencies by the length of the CDR3, default is False}
}
\value{
a sparse matrix of sequences versus kmer counts
}
\description{
Count kmer frequencies in a CDR3
}
