% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{getPooledSamples}
\alias{getPooledSamples}
\title{Pool all repertoires}
\usage{
getPooledSamples(repSeqObj, g = NULL)
}
\arguments{
\item{repSeqObj}{is an object containing all repertoire sample data}

\item{g}{a vector indicating the group to sample from, default is NULL and pools all samples from all groups, to pool samples only in group 1, g should be c(1)}
}
\value{
a single resepse dataset that has all CDR3 sequences from all samples, the frequency of all CDR3s is updated depending o the original counts in the original data.
}
\description{
Pool all repertoires
}

