% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepAnFns.R
\name{getUnproductive}
\alias{getUnproductive}
\title{Get unproductive CDR3s from a repertoire data (Repseq sample) in immunoseq format}
\usage{
getUnproductive(rep)
}
\arguments{
\item{rep}{name of the repertoire sample (which has already been read in with for example with read.table)}
}
\value{
a subset of the repertoire (rep) containing only unproductive CDR3s is returned
}
\description{
Get unproductive CDR3s from a repertoire data (Repseq sample) in immunoseq format
}
