% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{item.exists}
\alias{item.exists}
\title{check if item exists in repseq object}
\usage{
item.exists(object, item)
}
\arguments{
\item{object}{is a repseq object containing all repertoire sample data}

\item{item}{is the name of the item checked if it exists in the repseq object}
}
\value{
true if the item exists in object, false if item does not exist in object
}
\description{
check if item exists in repseq object
}
