% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepAnFns.R
\name{loadBioconductorPacks}
\alias{loadBioconductorPacks}
\title{Attaches packages from Bioconductor after installing them if they don't exist already}
\usage{
loadBioconductorPacks(package.list = c("Biostrings", "RankProd",
  "preprocessCore", "msa", "ggseqlogo"))
}
\arguments{
\item{package.list}{vector of Bioconductor packages}
}
\description{
Attaches packages from Bioconductor after installing them if they don't exist already
}
