% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepAnFns.R
\name{loadPacks}
\alias{loadPacks}
\title{Attaches packages from CRAN after installing them if they don't exist already}
\usage{
loadPacks(package.list = c("ggplot2", "seqRFLP", "stringr", "permute",
  "cluster", "data.table", "NbClust", "doParallel", "dendextend", "kmer",
  "ape", "Matrix", "wordspace", "dynamicTreeCut", "e1071", "fclust",
  "randomForest", "preprocessCore", "gplots"))
}
\arguments{
\item{package.list}{vector of packages}
}
\description{
Attaches packages from CRAN after installing them if they don't exist already
}
