% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepAnFns.R
\name{readSample}
\alias{readSample}
\title{Read in Repertoire data (Repseq sample) in older Immunoseq format}
\usage{
readSample(samplename, cloneStatus = "p")
}
\arguments{
\item{samplename}{name of the immunoseq formatted repertoire sample (full address or in the working directory)}

\item{cloneStatus}{logical; reading only productive or only unproductive CDR3s from the sample; default is productive only}
}
\value{
a repseq sample is returned
}
\description{
Read in Repertoire data (Repseq sample) in older Immunoseq format
}
