\name{RepAn-package}
\alias{RepAn-package}
\alias{RepAn}
\docType{package}
\title{
\packageTitle{RepAn}
}
\description{
\packageDescription{RepAn}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RepAn}
\packageIndices{RepAn}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{RepAn}

Maintainer: \packageMaintainer{RepAn}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
