% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{resampleReferenceSamples}
\alias{resampleReferenceSamples}
\title{Pool and resample samples to create reference/background repertoires, normalize samples to equal number of sequence reads}
\usage{
resampleReferenceSamples(repSeqObj, totalReads = 1e+05, resample = T,
  resampleSize = 1000, useProb = T)
}
\arguments{
\item{repSeqObj}{is an object containing all repertoire sample data}

\item{totalReads}{the total size of reads per sample to scale our data to. All repertoires are scaled to 100k reads by default.}

\item{resample}{a boolean argument to decide whether to resample repertoires, needed when data is too big, default is True. If False, resamples as big as the original datasets are drawn from the pooled repertoire dataset}

\item{resampleSize}{Maximum possible number of clonotype reads repertoires will be resampled to, default is 1000.}

\item{useProb}{a boolean to indicate whether the resampling should be done probablistically or not.}
}
\value{
repSeq data object created by downsampling after pooling all datasets
}
\description{
Pool and resample samples to create reference/background repertoires, normalize samples to equal number of sequence reads
}

