% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{runDaAnalysis}
\alias{runDaAnalysis}
\title{This function performs differential abundance analysis between groups of TCRB CDR3 samples (repseq data) to identify differentially abundant (DA) CDR3s.}
\usage{
runDaAnalysis(repSeqObj, clusterby = "NT", kmerWidth = 4, paired = T,
  clusterDaPcutoff = 0.1, positionWt = F, distMethod = c("euclidean",
  "cosine"), useDynamicTreeCut = T, matchingMethod = "km",
  repeatResample = T, nRepeats = 10, resampleSize = 5000,
  useProb = T, returnAll = T, nRR = 1000)
}
\arguments{
\item{repSeqObj}{is an object containing all repertoire sample data}

\item{clusterby}{boolean; subsequence type to consider, either NT (nucleotide) or AA (amino acid), default is NT}

\item{kmerWidth}{subsequence width to use, default is 4 for NT, and 3 for AA clusterby}

\item{paired}{boolean; whether to perform paired analysis for matched datasets,default is true.}

\item{clusterDaPcutoff}{sub-repertoire level differential abundance testing cut off, default is 0.1. This works well for our test cases.}

\item{positionWt}{boolean; whether to use positional weights for kmer frequencies, default is false}

\item{distMethod}{the distance method to be used for distance calculation between CDR3 feature vectors, use "euclidean" for nt 4-mer, and "cosine" for aa 3-mer feature vectors}

\item{useDynamicTreeCut}{boolean; default true, uses Dynamic Tree cut algorithm to cut clustering dendrograms. if false, findOptimalK will be used to find optimal k}

\item{matchingMethod}{matching method to match cluster centroids from all samples to identify subrepertoires; default is km (kmeans). If hc, hierarchical clustering will be used with 
dynamic tree cut to define clusters, if og an in house algorithm will be used that matches each cluster centroid in first sample to their closest centroids in all samples.}

\item{repeatResample}{boolean; perform repeat resampling, default is true. If false, all repertoire dataset will be used for analysis without downsampling.}

\item{nRepeats}{number of repeat resample runs to perform if repeatResample is true, default is 10}

\item{resampleSize}{the downsampling size in the repeat resample runs. default is 5000}

\item{useProb}{boolean; if true, probabilistic sampling is performed for downsampling with most frequenty CDR3s being more likely to be resampled. If false, all CDR3s have equal chance of being resampled. Default is true.}

\item{returnAll}{boolean; if true, the function returns a list whose first element is a data frame with information on the candidate differentially abundant CDR3s and second element is the intermediate results for all repeat resamples. 
If false, the intermediate results fro all runs are not returned, only a data frame with candidate differentially abundant CDR3s and their filterig results is returned.}

\item{nRR}{the number of permutations to perform in the ranking step of candidate DA CDR3s to determine statistical significance.}
}
\value{
a data frame with all candidate DA CDR3s if returnAll is false, a list with data frame of candidate DA CDR3s and all intermediate results if returnAll is true.
}
\description{
Function performs clustering based differential abundance analysis of CDR3 sequences in two sample groups with repeat resampling strategy. It first
performs within sample unsupervised clustering using subsequence frequency based unsupervised clustering, matches the clusters to their closest match across samples, and performs differential abundance testing at the level of matching
clusters to identify differentially abundant condition associated CDR3 sequences
}
