% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{sampleFromPooled}
\alias{sampleFromPooled}
\title{Pool and resample samples to create reference/background repertoires and return only CDR3s, normalizes samples to equal number of sequence reads}
\usage{
sampleFromPooled(repSeqObj, nClones, cType = "AA", useProb = T)
}
\arguments{
\item{repSeqObj}{is an object containing all repertoire sample data}

\item{nClones}{number of CDR3 sequences to randomly draw}

\item{cType}{the type of CDR3 sequence to return for the drawn clones, either AA or NT}

\item{useProb}{a boolean to indicate whether the resampling should be done probablistically or not.}
}
\value{
returns a vector of CDR3 sequences with size nClones that is randomly sampled from the pooled datasets of all repseq samples in repSeqObj
}
\description{
Pool and resample samples to create reference/background repertoires and return only CDR3s, normalizes samples to equal number of sequence reads
}

