% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{sampleWithFeatuersFromPooled}
\alias{sampleWithFeatuersFromPooled}
\title{Pool and resample samples to create reference/background repertoires, normalizes samples to equal number of sequence reads}
\usage{
sampleWithFeatuersFromPooled(repSeqObj, nClones, useProb = T)
}
\arguments{
\item{repSeqObj}{is an object containing all repertoire sample data}

\item{nClones}{number of CDR3 sequences to randomly draw}

\item{useProb}{a boolean to indicate whether the resampling should be done probablistically or not.}
}
\value{
returns a single repseq data of randomly selected CDR3 sequences and their associated features, that is randomly sampled from the pooled datasets of all repseq samples in repSeqObj
}
\description{
Pool and resample samples to create reference/background repertoires, normalizes samples to equal number of sequence reads
}
