% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{scaleSamples}
\alias{scaleSamples}
\title{scale all samples to equal number of sequence reads}
\usage{
scaleSamples(repSeqObj, totalReads = 1e+05, resample = T,
  resampleSize = 3000, useProb = T)
}
\arguments{
\item{repSeqObj}{is a repseq data object containing all repertoire sample data}

\item{totalReads}{the total size of reads per sample to scale our data to. All repertoires are scaled to 100k reads by default.}

\item{resample}{a boolean argument to decide whether to resample repertoires, needed when data is too big, default is True}

\item{resampleSize}{Maximum possible number of clonotype reads repertoires will be resampled to, default is 1000.}

\item{useProb}{a boolean to indicate whether the resampling should be done probablistically or not.}
}
\value{
repSeq data Object containing normalized repertoire data after resampling
}
\description{
scale all samples to equal number of sequence reads
}

