% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{setUp}
\alias{setUp}
\title{Setup repseq object for differential abudance analysis, CDR3 counts are changed to counts per million for all samples}
\usage{
setUp(sampleNames = NULL, samGroup = NULL, normToCPM = T)
}
\arguments{
\item{sampleNames}{is a vector of sample names (that have been already read into R )}

\item{samGroup}{a vector of group labels, e.g c(0,0,1,1), 0 for samples in first condition and 1 for samples in second condition.}

\item{normToCPM}{normalize all repseq samples to a total repertoire size of 1e6 (1 million) by default.}
}
\value{
a repSeq object which is a list containing the raw data (normalized to 1e6 in all samples if normToCPM is set to true),the sample names, and their group status
}
\description{
Setup repseq object for differential abudance analysis, CDR3 counts are changed to counts per million for all samples
}
