% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDaAnalysisFns.R
\name{determineWeight}
\alias{determineWeight}
\title{Determine weight of kmer depending on position}
\usage{
determineWeight(kmer, seq)
}
\description{
Determine weight of kmer depending on position
}
\keyword{internal}
